affine: Remove unused variable rem from wide_int_constant_multiple_p
This might fix the current bootstrap failure on aarch64, I only tested it
on x86_64. But the rem variable is unused and the for poly_widest_int, there
could be loop if NUM_POLY_INT_COEFFS is 2 or more. In the case of aarch64,
NUM_POLY_INT_COEFFS is 2.
Note the reason why there is warning for the unused variable is due to the deconstructor.

Pushed as obvious after a build for x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-affine.cc (wide_int_constant_multiple_p): Remove unused rem variable.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>