c++: modules and local static
Here we weren't emitting the guard variable for 'a'  when we emitted 'afn'
in the importer, because we only treated inline variables as needing that.
Fixed by generalizing to vague_linkage_p.

But we need to specifically exempt vtables, because the rest of the module
code handles them specially and expects them to be DECL_EXTERNAL.

gcc/cp/ChangeLog:

	* module.cc (trees_out::core_bools): Check vague_linkage_p.
	(has_definition): Likewise.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/static-3_b.C: New test.
	* g++.dg/modules/static-3_a.H: New test.