aarch64: Fix fp8_scalar_1.c's stacktest1
The function body test was expecting:
	umov	w0, v0.b[0]
	strb	w0, [sp, 15]

But the code generation was improved after r15-5375-gbeec291225be to just:
	str	b0, [sp, 15]

which is correct and better because no longer need to move between SIMD registers
and the GPRs.
This changes the function body test to new better code generation.

Pushed as obvious after a test of the testcase to make sure it now passes.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/fp8_scalar_1.c (stacktest1): Fix for new
	improved code generation.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>