selftest: invoke "diff" when ASSERT_STREQ fails
Currently when ASSERT_STREQ or ASSERT_STREQ_AT fail we print
both strings to stderr.  However it can be hard to figure out
the problem (e.g. for 1-character differences in long strings).

Extend the output by writing out the strings to tempfiles and
invoking "diff -up" on them when we have such a selftest failure,
to (I hope) simplify debugging.

gcc/ChangeLog:
	* selftest.cc (selftest::print_diff): New function.
	(selftest::assert_streq): Call it when we have non-equal
	non-null strings.

Signed-off-by: David Malcolm <dmalcolm@redhat.com>