aarch64: Update error message check for __builtin_launder check of sve-sizeless-2.C
r15-3614-g9fe57e4879de93 changed the error message for __builtin_launder but this testcase
was not updated for the new format of the error message since it is an aarch64 specific
testcase.

This patch updates the expected error message.

Pushed as obvious after testing to see the testcase now works.

gcc/testsuite/ChangeLog:

	* g++.dg/ext/sve-sizeless-2.C: Update the expected error message
	for __builtin_launder.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>