ifcombine: skip fallback conjunction on noncontiguous blocks
When everything else fails, if enabled by the target or by a
parameter, and when other requirements are satisfied, ifcombine
generates an AND of both conditions.

That may be good for contiguous conditions, but it's unlikely to be an
optimization when the blocks are separate.

Add contiguity to the set of requirements for this fallback
transformation.


for  gcc/ChangeLog

	* tree-ssa-ifcombine.cc (ifcombine_ifandif): Avoid fallback
	conjunction of noncontiguous conditions.