c: Do not remove _Atomic from array element type for typeof_unqual [PR117781]
As reported in bug 117781, my fix for bug 112841 broke the case of
typeof_unqual applied to an array of _Atomic elements, which should
not have _Atomic removed since only the element type is atomic, not
the array type.  Fix with logic to ensure that atomic element types
are preserved as such, while other qualifiers (i.e. those that are
semantically rather than only syntactically such in C) are removed.

Bootstrapped with no regressions for x86_64-pc-linux-gnu.

	PR c/117781

gcc/c/
	* c-parser.cc (c_parser_typeof_specifier): Do not remove _Atomic
	from array element type for typeof_unqual.

gcc/testsuite/
	* gcc.dg/c23-typeof-5.c: New test.