c: Fix gimplification ICE for shifts with invalid redeclarations
As reported in bug 117757, there is a C gimplification ICE for shifts
involving a variable that was incompatibly redeclared (and thus had
its type changed to error_mark_node).  Fix this with an appropriate
error_operand_p check.

Note that this is not the same issue as any of the other bugs reported
for ICEs later in the gimplifier dealing with such erroneous
redeclarations (it is, however, the same as the *second* ICE reported
in bug 115644 - the test in comment#1 for that bug, not the one in the
original bug report).

Bootstrapped with no regressions for x86_64-pc-linux-gnu.

	PR c/117757

gcc/c-family/
	* c-gimplify.cc (c_gimplify_expr): Check for error_operand_p
	before calling TYPE_MAIN_VARIANT for shifts.

gcc/testsuite/
	* gcc.dg/pr117757-1.c: New test.