ada: Elide copy for calls in allocators for nonlimited by-reference types
This prevents a temporary from being created on the primary stack to hold
the result of the function calls before it is copied to the newly allocated
memory in the nonlimited by-reference case.

That's already not done in the nonlimited non-by-reference case and there is
no reason to do it in the former case either.  The main issue is the call to
Remove_Side_Effects in Expand_Allocator_Expression, but its only purpose is
to cover the problematic processing done in Build_Allocate_Deallocate_Proc
on (part of) the expression; once this is fixed, the call is unnecessary.

The change also contains another small fix to deal with the corner case of
allocators for access-to-access types.

gcc/ada/ChangeLog:

	* exp_ch4.adb (Expand_Allocator_Expression): Do not preventively
	call Remove_Side_Effects on the expression in the nonlimited
	by-reference case.  Always call Build_Allocate_Deallocate_Proc
	in the default case.
	* exp_ch6.adb (Expand_Ctrl_Function_Call): Bail out if the call
	is the qualified expression of an allocator.
	* exp_util.adb (Build_Allocate_Deallocate_Proc): Replace all the
	calls to Relocate_Node by calls to Duplicate_Subexpr_No_Checks.